requirements.txt
Flask==1.1.1
flask-heroku==0.1.9
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
psycopg2-binary==2.8.4
netCDF4==1.5.3
python-dotenv==0.12.0
